transcription/requirements_38_10.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
appdirs=1.4.4=pypi_0
audioread=2.1.9=pypi_0
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.10.8=py38h06a4308_2
cffi=1.15.0=pypi_0
charset-normalizer=2.0.11=pypi_0
ci-sdr=0.0.0=pypi_0
click=8.0.3=pypi_0
configargparse=1.5.3=pypi_0
ctc-segmentation=1.7.1=pypi_0
cython=0.29.27=pypi_0
decorator=5.1.1=pypi_0
distance=0.1.3=pypi_0
einops=0.4.0=pypi_0
espnet=0.10.6=pypi_0
espnet-model-zoo=0.1.7=pypi_0
espnet-tts-frontend=0.0.3=pypi_0
ffmpeg=1.4=pypi_0
filelock=3.4.2=pypi_0
g2p-en=2.1.0=pypi_0
h5py=3.6.0=pypi_0
huggingface-hub=0.4.0=pypi_0
humanfriendly=10.0=pypi_0
idna=3.3=pypi_0
inflect=5.4.0=pypi_0
jaconv=0.3=pypi_0
jamo=0.4.1=pypi_0
joblib=1.1.0=pypi_0
kaldiio=2.17.2=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
librosa=0.9.0=pypi_0
libstdcxx-ng=9.3.0=hd4cf53a_17
llvmlite=0.38.0=pypi_0
ncurses=6.3=h7f8727e_2
nltk=3.7=pypi_0
numba=0.55.1=pypi_0
numpy=1.21.5=pypi_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.4.0=pypi_0
pip=21.2.4=py38h06a4308_0
pooch=1.6.0=pypi_0
pycparser=2.21=pypi_0
pyparsing=3.0.7=pypi_0
pypinyin=0.44.0=pypi_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytorch-wpe=0.0.1=pypi_0
pytz=2021.3=pypi_0
pyworld=0.3.0=pypi_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.1.18=pypi_0
requests=2.27.1=pypi_0
resampy=0.2.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=58.0.4=py38h06a4308_0
six=1.16.0=pypi_0
soundfile=0.10.3.post1=pypi_0
sqlite=3.37.2=hc218d9a_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
torch=1.10.2=pypi_0
torch-complex=0.2.1=pypi_0
tqdm=4.62.3=pypi_0
typeguard=2.13.3=pypi_0
typing-extensions=4.0.1=pypi_0
unidecode=1.3.2=pypi_0
urllib3=1.26.8=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7f8727e_4
